Remove 'Naming', 'Organization', and 'Build'
This document should be limited to issues of style. Naming,
organization, and build processes are not strictly style but
architecture decisions.

Naming of HTML classes should live in a style guide for HTML.

Organization decisions should be part of an HTML/CSS architecture
document.

Build and deployment do not belong in a style guide.